refactor(folder structure) : make core folder
move util files & another class to core folder
