+++ src/core/Utils.cpp
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "buffer/Library.hpp"
+#include "../buffer/Library.hpp"
 
 void setNonblock(int fd) {
 	if (fcntl(fd, F_SETFL, O_NONBLOCK) < 0) {
