+++ src/core/WebservUtils.cpp
