+++ src/server/Server.hpp
@@ -17,7 +17,7 @@
 #include <sys/event.h>
 #include <sys/time.h>
 #include "ServerConfig.hpp"
-#include "../Webserv.hpp"
+#include "../core/Webserv.hpp"
 #include "../buffer/File.hpp"
 #include "../buffer/Message.hpp"
 #include "../buffer/Pipe.hpp"
